Add MailCache type
Which can be used to do certain operations on _sets_ of mails.

This can later be used to (for example) update the mail database by
verifying all mail refs and re-finding (or invalidate/delete) broken
references.

It can also be used to create links (in sense of libimaglink) between
references to mails based on the mails Message-Id or other predicates
(for example same author).